api/requirements.txt
asgiref==3.7.2
Django==4.2.10
sqlparse==0.4.4
psycopg[binary]==3.1.12
django-environ==0.8.0
djangorestframework==3.14.0
coverage==6.3.2
django-cors-headers==3.8.0
djangorestframework-simplejwt==5.0.0
